Make it possible to set the version from the outside
Package managers usually work on an exported source directory (i.e. not on git
repo), however vrt relies on git describe to guess the version that it is
building. This change makes it possible to set the SEMVER variable from outside
of the makefile, so package managers can avoid the dependency on git